+++ bmc-monitoring/src/main/java/com/oracle/bmc/monitoring/MonitoringAsync.java
@@ -59,7 +59,9 @@ public interface MonitoringAsync extends AutoCloseable {
      * Creates a new alarm in the specified compartment.
      * For important limits information, see [Limits on Monitoring](https://docs.cloud.oracle.com/iaas/Content/Monitoring/Concepts/monitoringoverview.htm#Limits).
      * <p>
-     * Transactions Per Second (TPS) per-tenancy limit for this operation: 1.
+     * This call is subject to a Monitoring limit that applies to the total number of requests across all alarm operations.
+     * Monitoring might throttle this call to reject an otherwise valid request when the total rate of alarm operations exceeds 10 requests,
+     * or transactions, per second (TPS) for a given tenancy.
      *
      *
      * @param request The request object containing the details to send
@@ -77,7 +79,9 @@ public interface MonitoringAsync extends AutoCloseable {
      * Deletes the specified alarm.
      * For important limits information, see [Limits on Monitoring](https://docs.cloud.oracle.com/iaas/Content/Monitoring/Concepts/monitoringoverview.htm#Limits).
      * <p>
-     * Transactions Per Second (TPS) per-tenancy limit for this operation: 1.
+     * This call is subject to a Monitoring limit that applies to the total number of requests across all alarm operations.
+     * Monitoring might throttle this call to reject an otherwise valid request when the total rate of alarm operations exceeds 10 requests,
+     * or transactions, per second (TPS) for a given tenancy.
      *
      *
      * @param request The request object containing the details to send
@@ -95,7 +99,9 @@ public interface MonitoringAsync extends AutoCloseable {
      * Gets the specified alarm.
      * For important limits information, see [Limits on Monitoring](https://docs.cloud.oracle.com/iaas/Content/Monitoring/Concepts/monitoringoverview.htm#Limits).
      * <p>
-     * Transactions Per Second (TPS) per-tenancy limit for this operation: 1.
+     * This call is subject to a Monitoring limit that applies to the total number of requests across all alarm operations.
+     * Monitoring might throttle this call to reject an otherwise valid request when the total rate of alarm operations exceeds 10 requests,
+     * or transactions, per second (TPS) for a given tenancy.
      *
      *
      * @param request The request object containing the details to send
@@ -113,7 +119,9 @@ public interface MonitoringAsync extends AutoCloseable {
      * Get the history of the specified alarm.
      * For important limits information, see [Limits on Monitoring](https://docs.cloud.oracle.com/iaas/Content/Monitoring/Concepts/monitoringoverview.htm#Limits).
      * <p>
-     * Transactions Per Second (TPS) per-tenancy limit for this operation: 1.
+     * This call is subject to a Monitoring limit that applies to the total number of requests across all alarm operations.
+     * Monitoring might throttle this call to reject an otherwise valid request when the total rate of alarm operations exceeds 10 requests,
+     * or transactions, per second (TPS) for a given tenancy.
      *
      *
      * @param request The request object containing the details to send
@@ -132,7 +140,9 @@ public interface MonitoringAsync extends AutoCloseable {
      * Lists the alarms for the specified compartment.
      * For important limits information, see [Limits on Monitoring](https://docs.cloud.oracle.com/iaas/Content/Monitoring/Concepts/monitoringoverview.htm#Limits).
      * <p>
-     * Transactions Per Second (TPS) per-tenancy limit for this operation: 1.
+     * This call is subject to a Monitoring limit that applies to the total number of requests across all alarm operations.
+     * Monitoring might throttle this call to reject an otherwise valid request when the total rate of alarm operations exceeds 10 requests,
+     * or transactions, per second (TPS) for a given tenancy.
      *
      *
      * @param request The request object containing the details to send
@@ -150,7 +160,9 @@ public interface MonitoringAsync extends AutoCloseable {
      * List the status of each alarm in the specified compartment.
      * For important limits information, see [Limits on Monitoring](https://docs.cloud.oracle.com/iaas/Content/Monitoring/Concepts/monitoringoverview.htm#Limits).
      * <p>
-     * Transactions Per Second (TPS) per-tenancy limit for this operation: 1.
+     * This call is subject to a Monitoring limit that applies to the total number of requests across all alarm operations.
+     * Monitoring might throttle this call to reject an otherwise valid request when the total rate of alarm operations exceeds 10 requests,
+     * or transactions, per second (TPS) for a given tenancy.
      *
      *
      * @param request The request object containing the details to send
@@ -170,7 +182,7 @@ public interface MonitoringAsync extends AutoCloseable {
      * For information about metrics, see [Metrics Overview](https://docs.cloud.oracle.com/iaas/Content/Monitoring/Concepts/monitoringoverview.htm#MetricsOverview).
      * For important limits information, see [Limits on Monitoring](https://docs.cloud.oracle.com/iaas/Content/Monitoring/Concepts/monitoringoverview.htm#Limits).
      * <p>
-     * Transactions Per Second (TPS) per-tenancy limit for this operation: 1.
+     * Transactions Per Second (TPS) per-tenancy limit for this operation: 10.
      *
      *
      * @param request The request object containing the details to send
@@ -221,7 +233,9 @@ public interface MonitoringAsync extends AutoCloseable {
      * Removes any existing suppression for the specified alarm.
      * For important limits information, see [Limits on Monitoring](https://docs.cloud.oracle.com/iaas/Content/Monitoring/Concepts/monitoringoverview.htm#Limits).
      * <p>
-     * Transactions Per Second (TPS) per-tenancy limit for this operation: 1.
+     * This call is subject to a Monitoring limit that applies to the total number of requests across all alarm operations.
+     * Monitoring might throttle this call to reject an otherwise valid request when the total rate of alarm operations exceeds 10 requests,
+     * or transactions, per second (TPS) for a given tenancy.
      *
      *
      * @param request The request object containing the details to send
@@ -262,7 +276,9 @@ public interface MonitoringAsync extends AutoCloseable {
      * Updates the specified alarm.
      * For important limits information, see [Limits on Monitoring](https://docs.cloud.oracle.com/iaas/Content/Monitoring/Concepts/monitoringoverview.htm#Limits).
      * <p>
-     * Transactions Per Second (TPS) per-tenancy limit for this operation: 1.
+     * This call is subject to a Monitoring limit that applies to the total number of requests across all alarm operations.
+     * Monitoring might throttle this call to reject an otherwise valid request when the total rate of alarm operations exceeds 10 requests,
+     * or transactions, per second (TPS) for a given tenancy.
      *
      *
      * @param request The request object containing the details to send