+++ bmc-objectstorage/bmc-objectstorage-generated/src/main/java/com/oracle/bmc/objectstorage/ObjectStorage.java
@@ -103,8 +103,9 @@ public interface ObjectStorage extends AutoCloseable {
 
     /**
      * Deletes a bucket if the bucket is already empty. If the bucket is not empty, use
-     * {@link #deleteObject(DeleteObjectRequest) deleteObject} first. You also cannot
-     * delete a bucket that has a pre-authenticated request associated with that bucket.
+     * {@link #deleteObject(DeleteObjectRequest) deleteObject} first. In addition,
+     * you cannot delete a bucket that has a multipart upload in progress or a pre-authenticated
+     * request associated with that bucket.
      *
      * @param request The request object containing the details to send
      * @return A response object containing details about the completed operation
@@ -168,7 +169,7 @@ public interface ObjectStorage extends AutoCloseable {
      * Gets the metadata for the Object Storage namespace, which contains defaultS3CompartmentId and
      * defaultSwiftCompartmentId.
      * <p>
-     * Any user with the NAMESPACE_READ permission will be able to see the current metadata. If you are
+     * Any user with the OBJECTSTORAGE_NAMESPACE_READ permission will be able to see the current metadata. If you are
      * not authorized, talk to an administrator. If you are an administrator who needs to write policies
      * to give users access, see
      * [Getting Started with Policies](https://docs.cloud.oracle.com/Content/Identity/Concepts/policygetstarted.htm).
@@ -335,6 +336,23 @@ public interface ObjectStorage extends AutoCloseable {
             PutObjectLifecyclePolicyRequest request);
 
     /**
+     * Reencrypts the data encryption key of the bucket and objects in the bucket. This is an asynchronous call, the
+     * system will start a work request task to reencrypt the data encryption key of the objects and chunks in the bucket.
+     * Only the objects created before the time the API call will be reencrypted. The call can take long time depending
+     * on how many objects in the bucket and how big the objects are. This API will return a work request id, so the user
+     * can use this id to retrieve the status of the work request task.
+     * <p>
+     * A user can update kmsKeyId of the bucket, and then call this API, so the data encryption key of the bucket and
+     * objects in the bucket will be reencryped by the new kmsKeyId. Note that the system doesn't maintain what
+     * ksmKeyId is used to encrypt the object, the user has to maintain the mapping if they want.
+     *
+     * @param request The request object containing the details to send
+     * @return A response object containing details about the completed operation
+     * @throws BmcException when an error occurs.
+     */
+    ReencryptBucketResponse reencryptBucket(ReencryptBucketRequest request);
+
+    /**
      * Rename an object in the given Object Storage namespace.
      *
      * @param request The request object containing the details to send
@@ -355,6 +373,10 @@ public interface ObjectStorage extends AutoCloseable {
 
     /**
      * Performs a partial or full update of a bucket's user-defined metadata.
+     * <p>
+     * Use UpdateBucket to move a bucket from one compartment to another within the same tenancy. Supply the compartmentID
+     * of the compartment that you want to move the bucket to. For more information about moving resources between compartments,
+     * see [Moving Resources to a Different Compartment](https://docs.cloud.oracle.com/iaas/Content/Identity/Tasks/managingcompartments.htm#moveRes).
      *
      * @param request The request object containing the details to send
      * @return A response object containing details about the completed operation
@@ -368,7 +390,7 @@ public interface ObjectStorage extends AutoCloseable {
      * <p>
      * You can change the default Swift/Amazon S3 compartmentId designation to a different compartmentId. All
      * subsequent bucket creations will use the new default compartment, but no previously created
-     * buckets will be modified. A user must have NAMESPACE_UPDATE permission to make changes to the default
+     * buckets will be modified. A user must have OBJECTSTORAGE_NAMESPACE_UPDATE permission to make changes to the default
      * compartments for Amazon S3 and Swift.
      *
      * @param request The request object containing the details to send