+++ bmc-objectstorage/bmc-objectstorage-generated/src/main/java/com/oracle/bmc/objectstorage/ObjectStorageClient.java
@@ -1270,6 +1270,35 @@ public class ObjectStorageClient implements ObjectStorage {
     }
 
     @Override
+    public ReencryptBucketResponse reencryptBucket(ReencryptBucketRequest request) {
+        LOG.trace("Called reencryptBucket");
+        final ReencryptBucketRequest interceptedRequest =
+                ReencryptBucketConverter.interceptRequest(request);
+        com.oracle.bmc.http.internal.WrappedInvocationBuilder ib =
+                ReencryptBucketConverter.fromRequest(client, interceptedRequest);
+        com.google.common.base.Function<javax.ws.rs.core.Response, ReencryptBucketResponse>
+                transformer = ReencryptBucketConverter.fromResponse();
+
+        final com.oracle.bmc.retrier.BmcGenericRetrier retrier =
+                com.oracle.bmc.retrier.Retriers.createPreferredRetrier(
+                        interceptedRequest.getRetryConfiguration(), retryConfiguration);
+        return retrier.execute(
+                interceptedRequest,
+                retryRequest -> {
+                    final com.oracle.bmc.retrier.TokenRefreshRetrier tokenRefreshRetrier =
+                            new com.oracle.bmc.retrier.TokenRefreshRetrier(
+                                    authenticationDetailsProvider);
+                    return tokenRefreshRetrier.execute(
+                            retryRequest,
+                            retriedRequest -> {
+                                javax.ws.rs.core.Response response =
+                                        client.post(ib, retriedRequest);
+                                return transformer.apply(response);
+                            });
+                });
+    }
+
+    @Override
     public RenameObjectResponse renameObject(RenameObjectRequest request) {
         LOG.trace("Called renameObject");
         final RenameObjectRequest interceptedRequest =