+++ bmc-objectstorage/bmc-objectstorage-generated/src/main/java/com/oracle/bmc/objectstorage/requests/ReencryptBucketRequest.java
@@ -1,49 +1,31 @@
 /**
  * Copyright (c) 2016, 2019, Oracle and/or its affiliates. All rights reserved.
  */
-package com.oracle.bmc.resourcemanager.requests;
+package com.oracle.bmc.objectstorage.requests;
 
-import com.oracle.bmc.resourcemanager.model.*;
+import com.oracle.bmc.objectstorage.model.*;
 
-@javax.annotation.Generated(value = "OracleSDKGenerator", comments = "API Version: 20180917")
+@javax.annotation.Generated(value = "OracleSDKGenerator", comments = "API Version: 20160918")
 @lombok.Builder(builderClassName = "Builder", buildMethodName = "buildWithoutInvocationCallback")
 @lombok.Getter
-public class ChangeStackCompartmentRequest extends com.oracle.bmc.requests.BmcRequest {
+public class ReencryptBucketRequest extends com.oracle.bmc.requests.BmcRequest {
 
     /**
-     * The stack OCID.
+     * The Object Storage namespace used for the request.
      */
-    private String stackId;
+    private String namespaceName;
 
     /**
-     * Defines the properties of changeStackCompartment operation.
-     */
-    private ChangeStackCompartmentDetails changeStackCompartmentDetails;
-
-    /**
-     * For optimistic concurrency control. In the `PUT` or `DELETE` call for a resource, set the `if-match`
-     * parameter to the value of the etag from a previous `GET` or `POST` response for that resource.  The resource
-     * will be updated or deleted only if the etag you provide matches the resource's current etag value.
+     * The name of the bucket. Avoid entering confidential information.
+     * Example: `my-new-bucket1`
      *
      */
-    private String ifMatch;
+    private String bucketName;
 
     /**
-     * Unique Oracle-assigned identifier for the request. If you need to contact Oracle about a
-     * particular request, please provide the request ID.
-     *
+     * The client request ID for tracing.
      */
-    private String opcRequestId;
-
-    /**
-     * A token that uniquely identifies a request so it can be retried in case of a timeout or
-     * server error without risk of retrying the same action. Retry tokens expire after
-     * 24 hours, but can be invalidated before then due to conflicting operations. For example,
-     * if a resource has been deleted and purged from the system, then a retry of the original
-     * creation request may be rejected.
-     *
-     */
-    private String opcRetryToken;
+    private String opcClientRequestId;
 
     public static class Builder {
         private com.oracle.bmc.util.internal.Consumer<javax.ws.rs.client.Invocation.Builder>
@@ -77,29 +59,27 @@ public class ChangeStackCompartmentRequest extends com.oracle.bmc.requests.BmcRe
          * Copy method to populate the builder with values from the given instance.
          * @return this builder instance
          */
-        public Builder copy(ChangeStackCompartmentRequest o) {
-            stackId(o.getStackId());
-            changeStackCompartmentDetails(o.getChangeStackCompartmentDetails());
-            ifMatch(o.getIfMatch());
-            opcRequestId(o.getOpcRequestId());
-            opcRetryToken(o.getOpcRetryToken());
+        public Builder copy(ReencryptBucketRequest o) {
+            namespaceName(o.getNamespaceName());
+            bucketName(o.getBucketName());
+            opcClientRequestId(o.getOpcClientRequestId());
             invocationCallback(o.getInvocationCallback());
             retryConfiguration(o.getRetryConfiguration());
             return this;
         }
 
         /**
-         * Build the instance of ChangeStackCompartmentRequest as configured by this builder
+         * Build the instance of ReencryptBucketRequest as configured by this builder
          *
          * Note that this method takes calls to {@link Builder#invocationCallback(com.oracle.bmc.util.internal.Consumer)} into account,
          * while the method {@link Builder#buildWithoutInvocationCallback} does not.
          *
          * This is the preferred method to build an instance.
          *
-         * @return instance of ChangeStackCompartmentRequest
+         * @return instance of ReencryptBucketRequest
          */
-        public ChangeStackCompartmentRequest build() {
-            ChangeStackCompartmentRequest request = buildWithoutInvocationCallback();
+        public ReencryptBucketRequest build() {
+            ReencryptBucketRequest request = buildWithoutInvocationCallback();
             request.setInvocationCallback(invocationCallback);
             request.setRetryConfiguration(retryConfiguration);
             return request;