+++ bmc-resourcemanager/src/main/java/com/oracle/bmc/resourcemanager/model/CreateStackDetails.java
@@ -69,6 +69,15 @@ public class CreateStackDetails {
             return this;
         }
 
+        @com.fasterxml.jackson.annotation.JsonProperty("terraformVersion")
+        private String terraformVersion;
+
+        public Builder terraformVersion(String terraformVersion) {
+            this.terraformVersion = terraformVersion;
+            this.__explicitlySet__.add("terraformVersion");
+            return this;
+        }
+
         @com.fasterxml.jackson.annotation.JsonProperty("freeformTags")
         private java.util.Map<String, String> freeformTags;
 
@@ -99,6 +108,7 @@ public class CreateStackDetails {
                             description,
                             configSource,
                             variables,
+                            terraformVersion,
                             freeformTags,
                             definedTags);
             __instance__.__explicitlySet__.addAll(__explicitlySet__);
@@ -113,6 +123,7 @@ public class CreateStackDetails {
                             .description(o.getDescription())
                             .configSource(o.getConfigSource())
                             .variables(o.getVariables())
+                            .terraformVersion(o.getTerraformVersion())
                             .freeformTags(o.getFreeformTags())
                             .definedTags(o.getDefinedTags());
 
@@ -158,6 +169,12 @@ public class CreateStackDetails {
      **/
     @com.fasterxml.jackson.annotation.JsonProperty("variables")
     java.util.Map<String, String> variables;
+
+    /**
+     * The stack's Terraform version
+     **/
+    @com.fasterxml.jackson.annotation.JsonProperty("terraformVersion")
+    String terraformVersion;
 
     /**
      * Free-form tags associated with this resource. Each tag is a simple key-value pair with no predefined name, type, or namespace.