+++ bmc-resourcemanager/src/main/java/com/oracle/bmc/resourcemanager/requests/ListTerraformVersionsRequest.java
@@ -8,25 +8,7 @@ import com.oracle.bmc.resourcemanager.model.*;
 @javax.annotation.Generated(value = "OracleSDKGenerator", comments = "API Version: 20180917")
 @lombok.Builder(builderClassName = "Builder", buildMethodName = "buildWithoutInvocationCallback")
 @lombok.Getter
-public class ChangeStackCompartmentRequest extends com.oracle.bmc.requests.BmcRequest {
-
-    /**
-     * The stack OCID.
-     */
-    private String stackId;
-
-    /**
-     * Defines the properties of changeStackCompartment operation.
-     */
-    private ChangeStackCompartmentDetails changeStackCompartmentDetails;
-
-    /**
-     * For optimistic concurrency control. In the `PUT` or `DELETE` call for a resource, set the `if-match`
-     * parameter to the value of the etag from a previous `GET` or `POST` response for that resource.  The resource
-     * will be updated or deleted only if the etag you provide matches the resource's current etag value.
-     *
-     */
-    private String ifMatch;
+public class ListTerraformVersionsRequest extends com.oracle.bmc.requests.BmcRequest {
 
     /**
      * Unique Oracle-assigned identifier for the request. If you need to contact Oracle about a
@@ -36,14 +18,9 @@ public class ChangeStackCompartmentRequest extends com.oracle.bmc.requests.BmcRe
     private String opcRequestId;
 
     /**
-     * A token that uniquely identifies a request so it can be retried in case of a timeout or
-     * server error without risk of retrying the same action. Retry tokens expire after
-     * 24 hours, but can be invalidated before then due to conflicting operations. For example,
-     * if a resource has been deleted and purged from the system, then a retry of the original
-     * creation request may be rejected.
-     *
+     * The compartment OCID on which to filter.
      */
-    private String opcRetryToken;
+    private String compartmentId;
 
     public static class Builder {
         private com.oracle.bmc.util.internal.Consumer<javax.ws.rs.client.Invocation.Builder>
@@ -77,29 +54,26 @@ public class ChangeStackCompartmentRequest extends com.oracle.bmc.requests.BmcRe
          * Copy method to populate the builder with values from the given instance.
          * @return this builder instance
          */
-        public Builder copy(ChangeStackCompartmentRequest o) {
-            stackId(o.getStackId());
-            changeStackCompartmentDetails(o.getChangeStackCompartmentDetails());
-            ifMatch(o.getIfMatch());
+        public Builder copy(ListTerraformVersionsRequest o) {
             opcRequestId(o.getOpcRequestId());
-            opcRetryToken(o.getOpcRetryToken());
+            compartmentId(o.getCompartmentId());
             invocationCallback(o.getInvocationCallback());
             retryConfiguration(o.getRetryConfiguration());
             return this;
         }
 
         /**
-         * Build the instance of ChangeStackCompartmentRequest as configured by this builder
+         * Build the instance of ListTerraformVersionsRequest as configured by this builder
          *
          * Note that this method takes calls to {@link Builder#invocationCallback(com.oracle.bmc.util.internal.Consumer)} into account,
          * while the method {@link Builder#buildWithoutInvocationCallback} does not.
          *
          * This is the preferred method to build an instance.
          *
-         * @return instance of ChangeStackCompartmentRequest
+         * @return instance of ListTerraformVersionsRequest
          */
-        public ChangeStackCompartmentRequest build() {
-            ChangeStackCompartmentRequest request = buildWithoutInvocationCallback();
+        public ListTerraformVersionsRequest build() {
+            ListTerraformVersionsRequest request = buildWithoutInvocationCallback();
             request.setInvocationCallback(invocationCallback);
             request.setRetryConfiguration(retryConfiguration);
             return request;